requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2015.9.6.2
funcsigs==0.4
kafka-python==0.9.4
mock==1.3.0
nose==1.3.7
numpy==1.10.1
oauthlib==1.0.3
pbr==1.8.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.6
redis==2.10.3
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
tornado==4.2.1
twython==3.3.0
ujson==1.33
wheel==0.24.0
